more tuningManager development
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonData/trunk@24112 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Load/lib/perl/TuningConfig/InternalTable.pm
@@ -382,8 +382,8 @@ SQL
     ApiCommonData::Load::TuningConfig::Log::addErrorLog("\n" . $dbh->errstr . "\n");
   }
 
-  # drop obsolete table, if we're doing that
-  if (defined $synonymRtn && $purgeObsoletes) {
+  # drop obsolete table, if we're doing that (and it exists)
+  if (defined $synonymRtn && $purgeObsoletes && $oldTable) {
     ApiCommonData::Load::TuningConfig::Log::addLog("    purging obsolete table " . $oldTable);    
     $dbh->do("drop table " . $oldTable)
       or ApiCommonData::Load::TuningConfig::Log::addErrorLog("\n" . $dbh->errstr . "\n");
